Create roller command and subsystem
The removal of the gripper means that a lot more files are now obsolete, so a ton of files have been moved.
