+++ src/main/java/frc/robot/Robot.java
@@ -4,10 +4,8 @@
 
 package frc.robot;
 
-import edu.wpi.first.networktables.NetworkTableEntry;
-import edu.wpi.first.networktables.NetworkTableInstance;
-import edu.wpi.first.wpilibj.PneumaticsModuleType;
-import edu.wpi.first.wpilibj.Solenoid;
+// import edu.wpi.first.wpilibj.PneumaticsModuleType;
+// import edu.wpi.first.wpilibj.Solenoid;
 import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
@@ -23,12 +21,12 @@ import frc.robot.sensors.UltrasonicAnalog;
 public class Robot extends TimedRobot {
   private Command m_autonomousCommand;
   private RobotContainer m_robotContainer;
-  private Solenoid m_ultrasonic_solenoid =
+  /*private Solenoid m_ultrasonic_solenoid =
       new Solenoid(
           Constants.CANDeviceIDs.kPCMID24V,
           PneumaticsModuleType.CTREPCM,
           Constants.UltrasonicConstants.kSensor_PCM_Channel);
-  private Solenoid m_trainning_soloenoid;
+  private Solenoid m_trainning_soloenoid;*/
   private DetectAprilTags detector = new DetectAprilTags();
 
   // UltrasonicAnalog sensor = new UltrasonicAnalog(Constants.GripperConstants.kSensorChannel);
@@ -43,7 +41,7 @@ public class Robot extends TimedRobot {
     m_robotContainer = RobotContainer.getInstance();
 
     // turn on the solenoid channel to power the ultrasonic sensor
-    m_ultrasonic_solenoid.set(true);
+    // m_ultrasonic_solenoid.set(true);
 
     // Starts up the camera and thread to detect april tags
     detector.initDetector();
@@ -116,12 +114,12 @@ public class Robot extends TimedRobot {
     // Cancels all running commands at the start of test mode.
     CommandScheduler.getInstance().cancelAll();
     // enable the solenoid channel to allow for training the ultrasonic sensor
-    m_trainning_soloenoid =
+    /*m_trainning_soloenoid =
         new Solenoid(
             Constants.CANDeviceIDs.kPCMID24V,
             PneumaticsModuleType.CTREPCM,
             Constants.UltrasonicConstants.kTrainSensor_PCM_Channel);
-    m_trainning_soloenoid.set(true);
+    m_trainning_soloenoid.set(true);*/
   }
 
   /** This function is called periodically during test mode. */
