+++ src/main/java/frc/robot/commands/manipulator/AlignGripperToDoubleSubstation.java
@@ -6,6 +6,7 @@ package frc.robot.commands.manipulator;
 
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import frc.robot.Constants.*;
+import frc.robot.obsolete.commands.manipulator.GripperOpen;
 
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
@@ -18,7 +19,8 @@ public class AlignGripperToDoubleSubstation extends ParallelCommandGroup {
     addCommands(
         new MoveLift(LiftConstants.kSubStation),
         new MoveReach(ReachConstants.kSubStation),
-        new MoveWrist(WristConstants.kSubStation),
-        new GripperOpen());
+        new MoveWrist(WristConstants.kSubStation)
+        //new GripperOpen()
+        );
   }
 }
