+++ src/main/java/frc/robot/commands/manipulator/ScoreCubePosition2.java
@@ -6,6 +6,7 @@ package frc.robot.commands.manipulator;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.Constants;
+import frc.robot.obsolete.commands.manipulator.CubeGripperClose;
 
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
