+++ src/main/java/frc/robot/obsolete/commands/AUTO_LeaveCommunityAndEngage.java
@@ -2,20 +2,20 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.manipulator;
+package frc.robot.obsolete.commands;
 
+// import com.kauailabs.navx.frc.AHRS;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import frc.robot.Constants;
+import frc.robot.subsystems.drivetrain.SwerveSubsystem;
 
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class GrabCube extends SequentialCommandGroup {
-  /** Creates a new GrabCube. */
-  public GrabCube() {
+public class AUTO_LeaveCommunityAndEngage extends SequentialCommandGroup {
+  /** Creates a new GroupSeqCom_MovePastLineAndBack. */
+  public AUTO_LeaveCommunityAndEngage(SwerveSubsystem drivetrain) {
     // Add your commands in the addCommands() call, e.g.
     // addCommands(new FooCommand(), new BarCommand());
-    addCommands(
-        new MoveWrist(Constants.WristConstants.kLowerSoftLimit) /* , new CubeGripperClose()*/);
+    addCommands(new MovePastLine(drivetrain), new EngageChargingStation(drivetrain));
   }
 }
