+++ src/main/java/frc/robot/obsolete/commands/CheckIfEngaged.java
@@ -1,30 +1,21 @@
 // Copyright (c) FIRST and other WPILib contributors.
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
-package frc.robot.commands.obsolete;
+
+package frc.robot.obsolete.commands;
 
 // import com.kauailabs.navx.frc.AHRS;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import frc.robot.commands.StayEngaged;
 import frc.robot.subsystems.drivetrain.SwerveSubsystem;
 
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class EngageChargingStation extends SequentialCommandGroup {
-
-  /** Creates a new EngageChargingStation. */
-  public EngageChargingStation(SwerveSubsystem driveTrain) {
+public class CheckIfEngaged extends SequentialCommandGroup {
+  /** Creates a new CheckIfEngaged. */
+  public CheckIfEngaged(SwerveSubsystem driveTrain) {
     // Add your commands in the addCommands() call, e.g.
     // addCommands(new FooCommand(), new BarCommand());
-    /*
-     * Face perpendicular to the charging station //Assume already true
-     * Drive forwards onto the charging station
-     * Stop once the charging station becomes level
-     */
-    addCommands(
-        new GetOnChargingStation(driveTrain), // Drive until pitch changes
-        new StayEngaged(driveTrain));
-    // new StopDrive(driveTrain));
+    addCommands(new IfClimbing(driveTrain), new IfLevel(driveTrain));
   }
 }
