+++ src/main/java/frc/robot/obsolete/commands/IfClimbing.java
@@ -2,19 +2,21 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.manipulator;
+package frc.robot.obsolete.commands;
 
+// import com.kauailabs.navx.frc.AHRS;
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.manipulator.Gripper;
+import frc.robot.Constants;
+import frc.robot.subsystems.drivetrain.SwerveSubsystem;
 
-public class GripperOpen extends CommandBase {
-  /** Creates a new OpenManipulator. */
-  Gripper m_gripper;
+public class IfClimbing extends CommandBase {
+  SwerveSubsystem m_drivetrain;
 
-  public GripperOpen() {
+  /** Creates a new IfClimbing. */
+  public IfClimbing(SwerveSubsystem driveTrain) {
     // Use addRequirements() here to declare subsystem dependencies.
-    m_gripper = Gripper.getInstance();
-    addRequirements(m_gripper);
+    m_drivetrain = driveTrain;
+    // addRequirements(driveTrain);
   }
 
   // Called when the command is initially scheduled.
@@ -23,9 +25,7 @@ public class GripperOpen extends CommandBase {
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() {
-    m_gripper.open();
-  }
+  public void execute() {}
 
   // Called once the command ends or is interrupted.
   @Override
@@ -34,6 +34,6 @@ public class GripperOpen extends CommandBase {
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return true;
+    return m_drivetrain.getPitch() >= Constants.PITCH_CLIMBING - Constants.PITCH_OFFSET;
   }
 }
