+++ src/main/java/frc/robot/obsolete/commands/IfLeavingStation.java
@@ -2,20 +2,18 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.obsolete;
+package frc.robot.obsolete.commands;
 
-// import com.kauailabs.navx.frc.AHRS;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Constants;
 import frc.robot.subsystems.drivetrain.SwerveSubsystem;
 
-public class IfClimbing extends CommandBase {
-  SwerveSubsystem m_drivetrain;
-
-  /** Creates a new IfClimbing. */
-  public IfClimbing(SwerveSubsystem driveTrain) {
+public class IfLeavingStation extends CommandBase {
+  SwerveSubsystem m_driveTrain;
+  /** Creates a new IfLeaving. */
+  public IfLeavingStation(SwerveSubsystem driveTrain) {
     // Use addRequirements() here to declare subsystem dependencies.
-    m_drivetrain = driveTrain;
+    m_driveTrain = driveTrain;
     // addRequirements(driveTrain);
   }
 
@@ -34,6 +32,6 @@ public class IfClimbing extends CommandBase {
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return m_drivetrain.getPitch() >= Constants.PITCH_CLIMBING - Constants.PITCH_OFFSET;
+    return m_driveTrain.getPitch() <= -Constants.PITCH_CLIMBING - Constants.PITCH_OFFSET;
   }
 }
