+++ src/main/java/frc/robot/obsolete/commands/manipulator/AutoGripperClose.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.manipulator;
+package frc.robot.obsolete.commands.manipulator;
 
 import frc.robot.Constants;
 import frc.robot.sensors.UltrasonicAnalog;
