+++ src/main/java/frc/robot/obsolete/commands/manipulator/CubeGripperOpen.java
@@ -2,19 +2,19 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.manipulator;
+package frc.robot.obsolete.commands.manipulator;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.manipulator.Gripper;
+import frc.robot.obsolete.subsystems.CubeGripper;
 
-public class GripperClose extends CommandBase {
-  /** Creates a new CloseManipulator. */
-  Gripper m_gripper;
+public class CubeGripperOpen extends CommandBase {
+  /** Creates a new OpenManipulator. */
+  CubeGripper m_cubegripper;
 
-  public GripperClose() {
+  public CubeGripperOpen() {
     // Use addRequirements() here to declare subsystem dependencies.
-    m_gripper = Gripper.getInstance();
-    addRequirements(m_gripper);
+    m_cubegripper = CubeGripper.getInstance();
+    addRequirements(m_cubegripper);
   }
 
   // Called when the command is initially scheduled.
@@ -23,13 +23,13 @@ public class GripperClose extends CommandBase {
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() {}
+  public void execute() {
+    m_cubegripper.open();
+  }
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {
-    m_gripper.close();
-  }
+  public void end(boolean interrupted) {}
 
   // Returns true when the command should end.
   @Override
