+++ src/main/java/frc/robot/obsolete/commands/manipulator/GrabCube.java
@@ -2,18 +2,21 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.manipulator;
+package frc.robot.obsolete.commands.manipulator;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
+import frc.robot.Constants;
+import frc.robot.commands.manipulator.MoveWrist;
 
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class ReleaseCube extends SequentialCommandGroup {
-  /** Creates a new ReleaseCube. */
-  public ReleaseCube() {
+public class GrabCube extends SequentialCommandGroup {
+  /** Creates a new GrabCube. */
+  public GrabCube() {
     // Add your commands in the addCommands() call, e.g.
     // addCommands(new FooCommand(), new BarCommand());
-    addCommands(new GripperOpen());
+    addCommands(
+        new MoveWrist(Constants.WristConstants.kLowerSoftLimit) /* , new CubeGripperClose()*/);
   }
 }
