+++ src/main/java/frc/robot/obsolete/commands/manipulator/GripperClose.java
@@ -2,19 +2,19 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.manipulator;
+package frc.robot.obsolete.commands.manipulator;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.manipulator.CubeGripper;
+import frc.robot.obsolete.subsystems.Gripper;
 
-public class CubeGripperClose extends CommandBase {
+public class GripperClose extends CommandBase {
   /** Creates a new CloseManipulator. */
-  CubeGripper m_cubegripper;
+  Gripper m_gripper;
 
-  public CubeGripperClose() {
+  public GripperClose() {
     // Use addRequirements() here to declare subsystem dependencies.
-    m_cubegripper = CubeGripper.getInstance();
-    addRequirements(m_cubegripper);
+    m_gripper = Gripper.getInstance();
+    addRequirements(m_gripper);
   }
 
   // Called when the command is initially scheduled.
@@ -23,13 +23,13 @@ public class CubeGripperClose extends CommandBase {
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() {
-    m_cubegripper.close();
-  }
+  public void execute() {}
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {}
+  public void end(boolean interrupted) {
+    m_gripper.close();
+  }
 
   // Returns true when the command should end.
   @Override
