+++ src/main/java/frc/robot/obsolete/commands/manipulator/ResetCubePosition.java
@@ -2,10 +2,13 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.manipulator;
+package frc.robot.obsolete.commands.manipulator;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.Constants;
+import frc.robot.commands.manipulator.MoveLift;
+import frc.robot.commands.manipulator.MoveReach;
+import frc.robot.commands.manipulator.MoveWrist;
 
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
